backend/requirements.txt
autogen_agentchat
openai
autogen
autogen_ext
pymongo
Flask
PyJWT
Werkzeug